Refactor forum functionality into it's own NgModule
