+++ src/app/app.module.ts
@@ -1,28 +1,18 @@
 import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
-import {HttpClientModule} from '@angular/common/http';
 
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
-import {ForumComponent} from './components/forum/forum.component';
-import {PostFeedComponent} from './components/post-feed/post-feed.component';
-import {PostComponent} from './components/post/post.component';
-import {AddPostComponent} from './components/add-post/add-post.component';
-import {ReactiveFormsModule} from '@angular/forms';
+import {ForumModule} from './forum/forum.module';
 
 @NgModule({
   declarations: [
-    AppComponent,
-    ForumComponent,
-    PostFeedComponent,
-    PostComponent,
-    AddPostComponent,
+    AppComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    HttpClientModule,
-    ReactiveFormsModule
+    ForumModule
   ],
   providers: [],
   bootstrap: [AppComponent]
