+++ src/app/forum/components/add-post/add-post.component.ts
