+++ src/app/forum/components/forum/forum.component.spec.ts
@@ -20,6 +20,6 @@ describe('ForumComponent', () => {
   });
 
   it('should create', () => {
-    expect(component).toBeTruthy();
+      expect(component).toBeTruthy();
   });
 });
