+++ src/app/forum/components/post-feed/post-feed.component.ts
@@ -15,7 +15,7 @@ export class PostFeedComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    this.postService.getPosts().subscribe(posts => this.posts = posts);
+      this.postService.getPosts().subscribe(posts => this.posts = posts);
   }
 
   addPost(post: Post) {
