+++ src/app/forum/components/post/post.component.ts
